Linear operators

Hamiltonian generation

The hamiltonian function can be used to generate a Chern insulator hamiltonian. The only essential parameter is m_lattice - it is a matrix with the size of the Chern insulator lattice which contains the value of the m parameter for each site. Its type must be CoordinateRepr. There is an alternative way to set the m_lattice parameter - if its type is a Matrix, an additional argument of type Symbol is required - a representation specifier (see Coordinate Representation for more detail).

Here is an example of both usages:

m_lattice = ones(15, 15)
m_lattice[6:10, 7:11] .= -1
m_repr = CoordinateRepr(m_lattice, :n)
H1 = hamiltonian(m_repr)
H2 = hamiltonian(m_lattice, :n)

H1 == H2 # These are equal ways to specify the m_lattice parameter

true

There is an ability to isolate some sites from other ones. To do that, a zone_mapping matrix is required. It is a CoordinateRepr{Symbol} object, which maps lattice sites to symbols, where different symbols mean different zones, and hoppings between sites in different zones are set to zero. You can define the mapping in the hamiltonian function call, or apply it to an existing hamiltonian matrix using the zones! function:

m_lattice = ones(15, 15)

mapping = CoordinateRepr(fill(:zone1, 15, 15))
mapping[6:10, 6:10] .= :zone2

H1 = hamiltonian(m_lattice, :n, zone_mapping = mapping)
H2 = hamiltonian(m_lattice, :n)
zones!(H2, mapping)

H1 == H2 # These are equal ways to specify the zone mapping

true

You also can apply some magnetic field to the Chern insulator. This can be done by passing a function that takes the position vector and returns the gauge vector potential. You can write your own function use one of the magnetic field macros defined:

@landau for Landau gauge field

@symm for symmetric gauge field

@flux for flux quantum

You can define the field in the hamiltonian function or apply it to an existing hamiltonian matrix using the field! function:

m_lattice = ones(15, 15)
B = 0.1

H0 = hamiltonian(m_lattice, :n, field = (x -> [0, x[1] * B, 0]))
H1 = hamiltonian(m_lattice, :n, field = @landau(B))
H2 = hamiltonian(m_lattice, :n)
field!(H2, @landau(B))

H0 == H1 == H2 # These are equal ways to specify the field

true

Other linear operators

To calculate a density matrix for zero temperature, you can use the filled_projector function. It accepts the hamiltonian matrix and returns a density matrix. You can also specify the Fermi level (which is zero by default).

You can also generate the coordinate operators using the coord_operators function. It takes an optional parameter symm which defines if the operators are defined symmetrically (i. e. the point with (0, 0) coordinates is in the center of the lattice) or not (i. e. the point with (0, 0) coordinates is in the bottom-left angle of the lattice).

Here is an example of usage of both functions to calculate the local Chern marker:

using Plots, LinearAlgebra

m_lattice = ones(15, 15)
m_lattice[6:10, 7:11] .= -1
H = hamiltonian(m_lattice, :n)

P = filled_projector(H)
X, Y = coord_operators()
c = 4pi * im * P * X * (I - P) * Y * P

heatmap(heatmap_data(c), color=:viridis)

Electric current

You can calculate electric currents using the currents function. It accepts the hamiltonian and the density matrix to produce a matrix with currents. Here is an example of electric currents in a magnetic field:

using Plots

B = 0.1
H = hamiltonian(ones(15, 15), :n, field=@symm(B))
P = filled_projector(H)
ps, qs = quiver_data(currents(H, P) * 10 / B)
quiver(ps, quiver = qs)

LCM current

There are several formulas which allow us to calculate currents of the local Chern marker. These formulas comply to the continuity equation, but, unlike electric currents, they are not localized. In this program, there are several macros that can interest you:

TopologicalMarkers.@J — Macro

@J(H, P, X, Y)

TopologicalMarkers.@J_best — Macro

@J_best(H, P, X, Y)

TopologicalMarkers.@J_c — Macro

@J_c(H, P, X, Y)

TopologicalMarkers.@J_eq — Macro

@J_eq(H, P, X, Y)

TopologicalMarkers.@J_inv — Macro

@J_inv(H, P, X, Y)

TopologicalMarkers.@J_m — Macro

@J_m(H, P, X, Y)

TopologicalMarkers.@J_m_inv — Macro

@J_m_inv(H, P, X, Y)

Each one takes matrices of the hamiltonian, the density and coordinate operators and returns a function that calculates a current given indices of 2 sites.

To generate a matrix with currents, use the @currents macro:

TopologicalMarkers.@currents — Macro

@currents(currents_lambda)

Generates a matrix with currents, given a lambda/macrocall that takes lattice site indices and returns the current value

Example usage:

currents_mat = @currents @J H P X Y