TopologicalMarkers.jl

TopologicalMarkers.CoordinateRepr

TopologicalMarkers.coord_operators

TopologicalMarkers.currents

TopologicalMarkers.domains!

TopologicalMarkers.evolution_operator

TopologicalMarkers.field!

TopologicalMarkers.filled_projector

TopologicalMarkers.hamiltonian

TopologicalMarkers.heatmap_data

TopologicalMarkers.optimal_layout

TopologicalMarkers.plot_auto

TopologicalMarkers.plot_boundaries!

TopologicalMarkers.plot_figure!

TopologicalMarkers.quiver_data

TopologicalMarkers.@J_b

TopologicalMarkers.@J_c

TopologicalMarkers.@J_eq

TopologicalMarkers.@J_inv

TopologicalMarkers.@J_m

TopologicalMarkers.@J_m_tr

TopologicalMarkers.@J_treq

TopologicalMarkers.@currents

TopologicalMarkers.@evolution

TopologicalMarkers.@flux

TopologicalMarkers.@landau

TopologicalMarkers.@symm

Hamiltonian matrix

TopologicalMarkers.hamiltonian — Function

hamiltonian{T}(m_repr; <keyword arguments>)
hamiltonian{T}(m_lattice, repr_spec; <keyword arguments>)

Generates a Hamiltonian operator for a Chern insulator using the following formula

\[\hat{H} = \sum_i m_i c^\dagger_i \sigma_z c_i + \sum_{x-links} c^\dagger_i \frac{\sigma_z - i \sigma_x}{2} c_j + \sum_{y-links} c^\dagger_i \frac{\sigma_z - i \sigma_y}{2} c_j + h. c.\]

Arguments

m_repr: The value of m on different sites, in CoordinateRepr format

Alternatively, pass a matrix and a representation specifier (see CoordinateRepr for more information)

Keyword arguments

pbc: Periodic boundary conditions. A Tuple{Bool, Bool}, each element sets boundary conditions for the horizontal and vertical edge respectively. Default is (false, false)

domains: A matrix with elements of arbitrary type, which maps sites to isolated domains. The hopping members between different domains are erased. There are no isolated domains by default

field: A function/lambda that takes two coordinates and returns the vector potential of the magnetic field. Used to calculate phase factors on hoppings. There is no magnetic field by default

TopologicalMarkers.field! — Function

field!(H, A[, lattice_size]; intervals = 10)

Applies magnetic field to specified hamiltonian. In other words, all hoppings get multiplied on a specific phase factor that can be calculated using Peierls substitution:

\[\varphi_{ij} = 2\pi \int_i^j A(r) \cdot dr\]

This integral is calculated explicitly for every hopping, using the A function.

Arguments

H: the hamiltonian matrix

A: a function that takes a Vector representing a point and returns a Vector representing the vector potential in that point

lattice_size: the size of the lattice the hamiltonian is defined for. If not provided, this function will use the value for the hamiltonian matrix that was created last

Keyword arguments

intervals: the number of intervals to use when calculating the Peierls substitution phase factor

TopologicalMarkers.domains! — Function

domains!(H, domain_mapping, repr_spec)
domains!(H, domain_mapping_repr)

Divides the Chern insulator hamiltonian into several unconnected domains. The hoppings between these domains are erased.

Arguments

H: the hamiltonian matrix

domain_mapping_repr: an object of type CoordinateRepr{Symbol}. Each site is mapped to a symbol, different symbols mean different domains

domain_mapping: domain mapping matrix. Each site is mapped to a symbol, different symbols mean different domains

repr_spec: representation specifier for the domain_mapping (see CoordinateRepr docs for more information)

TopologicalMarkers.@landau — Macro

@landau(B)

Generates a function that returns the Landau gauge vector potential. This can be used as an argument for the field! function.

Arguments

B: the magnetic field value

TopologicalMarkers.@symm — Macro

@symm(B[, center])

Generates a function that returns the symmetric gauge vector potential. This can be used as an argument for the field! function.

Arguments

B: the value of magnetic field

center: the center of symmetry of the vector potential

TopologicalMarkers.@flux — Macro

flux(Φ[, point])

Generates a function that returns the vector potential for a flux quantum. This can be used as an argument for the field! function.

Arguments

Φ: the value of magnetic field

point: the point where the flux is located

Other linear operators

TopologicalMarkers.coord_operators — Function

coord_operators([lattice_size]; symmetric = true)

Returns a tuple of coordinate operators (i. e. $\hat{X}$ and $\hat{Y}$).

Arguments

lattice_size: the size of the lattice

Keyword arguments

symmetric: true if the operator is symmetrically defined (in other words, the central site of the lattice corresponds to (0, 0)), false otherwise. True by default

TopologicalMarkers.filled_projector — Function

filled_projector(H; <keyword arguments>)

Returns a projector onto the filled states (in other words - a density matrix of the ground state).

Arguments

H: the hamiltonian matrix

Keyword arguments

fermi_energy: the Fermi energy level. 0 by default

T: temperature of the system, in energy units. 0 by default.

If value is non-zero, the state density will correspond to the Fermi-Dirac distribution

TopologicalMarkers.currents — Function

currents(H, P[, lattice_size])

Returns a skew-symmetric matrix of electric currents between sites.

Arguments

H: the hamiltonian matrix

P: the density matrix

lattice_size: the size of the lattice the hamiltonian is defined for

Bianca-Resta currents

TopologicalMarkers.@J_b — Macro

@J_b(H, P, X, Y)

Calculates the sum of @J_c and @J_m currents.

TopologicalMarkers.@J_c — Macro

@J_c(H, P, X, Y)

Calculates the Bianca-Resta current part using the following formula:

\[J_c(r, r') = \\ 4\pi ( \langle r | PXPYP | r' \rangle \langle r' | H | r \rangle - \langle r | H | r' \rangle \langle r' | PXPYP | r \rangle )\]

TopologicalMarkers.@J_eq — Macro

@J_eq(H, P, X, Y)

Calculates the Bianca-Resta current using the following formula:

\[J_{eq}(r, r') = \\ 4\pi ( \langle r | [P,H]XP | r' \rangle \langle r' | PYP | r \rangle - \langle r | PYP | r' \rangle \langle r' | [P,H]XP | r \rangle \\ - \langle r | [P,H]YP | r' \rangle \langle r' | PXP | r \rangle + \langle r | PXP | r' \rangle \langle r' | [P,H]YP | r \rangle )\]

TopologicalMarkers.@J_inv — Macro

@J_inv(H, P, X, Y)

Calculates the sum of @J_c and @J_m_inv currents.

TopologicalMarkers.@J_m — Macro

@J_m(H, P, X, Y)

Calculates the Bianca-Resta current part using the following formula:

\[J_m(r, r') = \\ 4\pi ( \langle r | PHXP | r' \rangle \langle r' | PYP | r \rangle - \langle r | PYP | r' \rangle \langle r' | PHXP | r \rangle \\ - \langle r | PHYP | r' \rangle \langle r' | PXP | r \rangle + \langle r | PXP | r' \rangle \langle r' | PHYP | r \rangle )\]

TopologicalMarkers.@J_m_tr — Macro

@J_m_tr(H, P, X, Y)

Calculates the Bianca-Resta current part using the following formula:

\[J_{m_{tr}}(r, r') = \\ 4\pi ( \langle r | P | r' \rangle \langle r' | [H,X] PYP | r \rangle - \langle r | [H,X] PYP | r' \rangle \langle r' | P | r \rangle \\ - \langle r | PYP | r' \rangle \langle r' | [H,X] P | r \rangle + \langle r | [H,X] P | r' \rangle \langle r' | PYP | r \rangle \\ - \langle r | P | r' \rangle \langle r' | [H,Y] PXP | r \rangle + \langle r | [H,Y] PXP | r' \rangle \langle r' | P | r \rangle \\ + \langle r | PXP | r' \rangle \langle r' | [H,Y] P | r \rangle - \langle r | [H,Y] P | r' \rangle \langle r' | PXP | r \rangle )\]

TopologicalMarkers.@J_treq — Macro

@J_treq(H, P, X, Y)

Calculates the Bianca-Resta current using the following formula:

\[J_{treq}(r, r') = \\ 4\pi ( \langle r | [H,P] P | r' \rangle \langle r' | PXPYP | r \rangle - \langle r | PXPYP | r' \rangle \langle r' | [H,P] P | r \rangle \\ - \langle r | [H,P] P | r' \rangle \langle r' | PYPXP | r \rangle + \langle r | PYPXP | r' \rangle \langle r' | [H,P] P | r \rangle \\ + \langle r | P | r' \rangle \langle r' | [H,P] XPYP | r \rangle - \langle r | [H,P] XPYP | r' \rangle \langle r' | P | r \rangle \\ + \langle r | P [H,P] XP | r' \rangle \langle r' | PYP | r \rangle - \langle r | PYP | r' \rangle \langle r' | P [H,P] XP | r \rangle \\ - \langle r | P | r' \rangle \langle r' | [H,P] YPXP | r \rangle + \langle r | [H,P] YPXP | r' \rangle \langle r' | P | r \rangle \\ - \langle r | P [H,P] YP | r' \rangle \langle r' | PXP | r \rangle + \langle r | PXP | r' \rangle \langle r' | P [H,P] YP | r \rangle )\]

TopologicalMarkers.@currents — Macro

@currents(currents_lambda)

Generates a matrix with currents, given a lambda/macro call that takes lattice site indices and returns the current value.

Example usage:

currents_mat = @currents @J_b H P X Y

Unitary evolution

TopologicalMarkers.evolution_operator — Function

evolution_operator(H, t)

Calculates the unitary evolution operator using the formula

$ \mathcal{U}(t) = e^{-\frac{1}{i\hbar} \hat{H} t} $

Arguments

H: the hamiltonian matrix

t: the evolution time

TopologicalMarkers.@evolution — Macro

@evolution [rules...] for_loop

Generates an environment with defined hamiltonian and density matrices that evolve by certain laws. See Unitary evolution for more details.

Data visualization

TopologicalMarkers.CoordinateRepr — Type

CoordinateRepr{T}

A wrapper struct for matrices to be conveniently plotted.

Arguments

lattice: A matrix representing some quantity defined on the lattice (e. g. the LCM)

repr_spec: A symbol defining the way how the lattice sites match the matrix values:

:c or :coord: The value for (x, y) site is A[x, y]. This is the default value

:n of :natural: If you print out the matrix as you usually do, and then imagine

a coordinate system with its center in the bottom-left corner, this will be the mapping between sites and matrix values

TopologicalMarkers.heatmap_data — Function

heatmap_data(op[, lattice_size])

Generates a CoordinateRepr for coordinate representation traces: $\langle r | \hat{\mathcal{O}} | r \rangle$.

Arguments

op: the operator to find values for

lattice_size: the size of the lattice

TopologicalMarkers.quiver_data — Function

quiver_data(currents_mat[, lattice_size]; <keyword arguments>)

Generates data for a quiver plot using a matrix with currents. The output is a tuple of two vectors with equal length: one contains arrow origins, the other one - arrow vectors.

Arguments

currents_mat: a matrix with currents

lattice_size: the size of the lattice

Keyword arguments

threshold: minimum value of the current to be put to output. Default is 0.1

dist_threshold: maximum distance between sites for which the current will be evaluated. Infinite by default

xlims and ylims: limit the area in which the currents will be evaluated. Infinite by default

Plotting shorthands

These functions may come in handy when you want to plot many heatmaps and electric current diagrams in one line.

TopologicalMarkers.plot_boundaries! — Function

plot_boundaries!([pl, ]domain_mapping; <keyword arguments>)

Draws boundaries between different domains described by the domain_mapping matrix.

Arguments

pl: a Plots.Plot object to visualize data on

domain_mapping: a CoordinateRepr object that represents domain mapping. The boundaries between different domains will be drawn

All keyword arguments will be passed to the plot! function used for drawing - this can be used to change the line thickness or style, for example.

TopologicalMarkers.plot_figure! — Function

plot_figure!(pl; <keyword arguments>)

Plots complicated splitline data series (heatmap, boundaries, quiver) on a single figure.

Arguments

pl: a Plots.Plot object to visualize data on

Keyword arguments

hmap: data to be visualized on a heatmap. It can be a CoordinateRepr object (then it will be plotted directly)

or a linear operator matrix (then the CoordinateRepr will be generated automatically)

domain_mapping: a CoordinateRepr object that represents domain mapping. The boundaries between different domains will be drawn.

currents: a matrix containing currents between sites

xlims and ylims: objects of type Tuple{Int, Int} that define the limits of the x- and y- axes respectively

All other keyword arguments with the following prefixes are passed to the plot! function:

hmap for the heatmap

bounds for the boundaries

currents for the quiver

This can be used to style the plot. For example, this line will result in the domain boundaries being drawn in dot style, the limits of the heatmap will be set from -3 to 3, and the arrows depicting the currents will be blue:

plot_figure!(..., hmapclims=(-3, 3), boundsstyle=:dot, :currentscolor=:green)

TopologicalMarkers.optimal_layout — Function

optimal_layout(figures_total; <keyword arguments>)

Generates optimal layout for multiple figures in a plot.

Arguments

figures_total: the number of figures to place on the plot

Keyword arguments

plot_aspect_ratio: The aspect ratio of the plot. 1:1 by default

figure_aspect_ratio: the aspect ratio of one figure of the plot. 7:5 by default

TopologicalMarkers.plot_auto — Function

plot_auto(<arguments>; <keyword arguments>)

Plots multiple heatmaps, cutaway views or currents simultaneously.

The subplots are automatically arranged into an optimal layout.

Arguments

Each argument can be either a CoordinateRepr object or a chain of pairs.

Keyword arguments

layout: a Plots.Layout object to use. If not specified, an optimal layout will be

generated depending on the plot size.

plot_size: actual size of the plot. If not specified, then the layout parameter will

be ignored and the size will be optimal for the layout

title: the title of the graph

cutaway_view: Tuple that indicates which cutaway view should be

drawn on a separate figure

cutaway_views: Same as previous, but many of them

All other keyword arguments with the following prefixes are passed to the plot! function:

hmap for the heatmap

bounds for the boundaries

currents for the quiver

cutaway for the cutaway figure

splitline for the line on all figures that shows the plane of the cutaway views