TopologicalMarkers.jl

TopologicalMarkers.CoordinateRepr

TopologicalMarkers.coord_operators

TopologicalMarkers.currents

TopologicalMarkers.evolution_operator

TopologicalMarkers.field!

TopologicalMarkers.filled_projector

TopologicalMarkers.hamiltonian

TopologicalMarkers.heatmap_data

TopologicalMarkers.plot_auto

TopologicalMarkers.plot_boundaries!

TopologicalMarkers.plot_marker!

TopologicalMarkers.quiver_data

TopologicalMarkers.zones!

TopologicalMarkers.@evolution

TopologicalMarkers.@flux

TopologicalMarkers.@landau

TopologicalMarkers.@symm

Hamiltonian matrix

TopologicalMarkers.hamiltonian — Function

hamiltonian{T}((m_repr | m_lattice, repr_spec); <keyword arguments>)

Generates a Hamiltonian operator for a Chern insulator using the following formula

\[\hat{H} = \sum_i m_i c^\dagger_i \sigma_z c_i + \sum_{x-links} c^\dagger_i \frac{\sigma_z - i \sigma_x}{2} c_j + \sum_{y-links} c^\dagger_i \frac{\sigma_z - i \sigma_y}{2} c_j + h. c.\]

Arguments

m_repr: The value of m on different sites, in CoordinateRepr format

Alternatively, pass a matrix and a representation specifier (see CoordinateRepr for more information)

pbc: Periodic boundary conditions. A Tuple{Bool, Bool}, each element sets boundary conditions for the horizontal and vertical edge respectively. Default is (false, false)

zones: A matrix with elements of arbitrary type, which maps sites to isolated zones. The hopping members between different zones are erased. There are no isolated zones by default

field: A function/lambda that takes two coordinates and returns the vector potential of the magnetic field. Used to calculate phase factors on hoppings. There is no magnetic field by default

TopologicalMarkers.field! — Function

field!(H, A[, lattice_size]; intervals = 10)

Applies magnetic field to specified hamiltonian. In other words, all hoppings get multiplied on a specific phase factor that can be calculated using Peierls substitution:

\[\varphi_{ij} = 2\pi \int_i^j A(r) \cdot dr\]

This integral is calculated explicitly for every hopping, using the A function.

Arguments

H: the hamiltonian matrix

A: a function that takes a Vector representing a point and returns a Vector representing the vector potential in that point

lattice_size: the size of the lattice the hamiltonian is defined for. If not provided, this function will use the value for the hamiltonian matrix that was created last

intervals: the number of intervals to use when calculating the Peierls substitution phase factor

TopologicalMarkers.zones! — Function

zones!(H, zone_mapping[, repr_spec])

Divides the Chern insulator hamiltonian into several unconnected zones. The hoppings between these zones are erased.

Arguments

H: the hamiltonian matrix

zone_mapping: an AbstractMatrix{Symbol} or CoordinateRepr{Symbol}. Each site is mapped to a symbol, different symbols mean different zones

repr_spec: if zone_mapping is an AbstractMatrix{Symbol}, this argument is a representation specifier (see CoordinateRepr docs for more information)

TopologicalMarkers.@landau — Macro

@landau(B)

Generates a function that returns the Landau gauge vector potential. This can be used as an argument for the field! function.

Arguments

B: the magnetic field value

TopologicalMarkers.@symm — Macro

@symm(B[, center])

Generates a function that returns the symmetric gauge vector potential. This can be used as an argument for the field! function.

Arguments

B: the value of magnetic field

center: the center of symmetry of the vector potential

TopologicalMarkers.@flux — Macro

flux(Φ[, point])

Generates a function that returns the vector potential for a flux quantum. This can be used as an argument for the field! function.

Arguments

Φ: the value of magnetic field

point: the point where the flux is located

Other linear operators

TopologicalMarkers.coord_operators — Function

coord_operators([lattice_size]; <keyword arguments>)

Returns a tuple of coordinate operators (e. g. $\hat{X}$ and $\hat{Y}$).

Arguments

lattice_size: the size of the lattice

symmetric: defines if the operators are symmetricall (e. g. the center of the lattice corresponds to (0, 0))

TopologicalMarkers.filled_projector — Function

filled_projector(H[, energy_thr=0])

Returns a projector onto the filled states (in other words, a ground state density matrix).

Arguments

H: the hamiltonian matrix

energy_thr: the Fermi energy level

TopologicalMarkers.currents — Function

currents(H, P[, lattice_size])

Returns a skew-symmetric matrix of electric currents between sites.

Arguments

H: the hamiltonian matrix.

P: the density matrix.

lattice_size: the size of the lattice the hamiltonian is defined for.

Unitary evolution

TopologicalMarkers.evolution_operator — Function

evolution_operator(H, t)

Calculates the unitary evolution operator using the formula

$ \mathcal{U}(t) = e^{-\frac{1}{i\hbar} \hat{H} t} $

Arguments

H: the hamiltonian matrix

t: the evolution time

TopologicalMarkers.@evolution — Macro

@evolution [rules...] for_loop

Generates an environment with defined hamiltonian and density matrices that evolve by certain laws. See Unitary evolution for more details.

Data visualization

TopologicalMarkers.CoordinateRepr — Type

CoordinateRepr{T}

A wrapper class for matrices to be conveniently plotted.

Arguments

lattice: A matrix representing some quantity defined on the lattice (e. g. the LCM)

repr_spec: A symbol defining the way how the lattice sites match the matrix values:

:c or :coord: The value for (x, y) site is A[x, y]. This is the default value

:n of :natural: If you print out the matrix as you usually do, and then imagine

a coordinate system with its center in the bottom-left corner, this will be the mapping between sites and matrix values

TopologicalMarkers.heatmap_data — Function

heatmap_data(op[, lattice_size])

Generates a CoordinateRepr for values like $\langle r | \hat{\mathcal{O}} | r \rangle$.

Arguments

op: the operator to find values for

lattice_size: the size of the lattice

TopologicalMarkers.quiver_data — Function

quiver_data(currents_mat[, lattice_size]; <keyword arguments>)

Generates data for a quiver plot using a matrix with currents. The output is a tuple of two vectors with equal length: one contains arrow origins, the other one - arrow vectors.

Arguments

currents_mat: a matrix with currents

lattice_size: the size of the lattice

threshold: minimum value of the current to be put to output. Default is 0.1

dist_threshold: maximum distance between sites for which the current will be evaluated. Infinite by default

xlims and ylims: limit the area in which the currents will be evaluated. Infinite by default

Plotting shorthands

These functions may come in handy when you want to plot many marker heatmaps and electric current diagrams in one line.

TopologicalMarkers.plot_boundaries! — Function

plot_boundaries!([pl, ]zone_mapping; <keyword arguments>)

Draws boundaries between different zones described by the zone_mapping matrix.

Arguments

pl: a Plots.Plot object to visualize data on

zone_mapping: a CoordinateRepr object that represents zone mapping. The boundaries between different zones will be drawn

All keyword arguments will be passed to the plot! function used for drawing - this can be used to change the line thickness or style, for example.

TopologicalMarkers.plot_marker! — Function

plot_marker!(pl; <keyword arguments>)

Plots complicated marker data series (heatmap, boundaries, quiver) on a single figure.

Arguments

pl: a Plots.Plot object to visualize data on

hmap: data to be visualized on a heatmap. It can be a CoordinateRepr object (then it will be plotted directly)

or a linear operator matrix (then the CoordinateRepr will be generated automatically)

zone_mapping: a CoordinateRepr object that represents zone mapping. The boundaries between different zones will be drawn.

currents: a matrix containing currents between sites

xlims and ylims: objects of type Tuple{Int, Int} that define the limits of the x- and y- axes respectively

All keyword arguments with different prefixes are passed th the plot! function:

hmap for the heatmap

bounds for the boundaries

currents for the quiver

This can be used to style the plot:

plot_marker!(..., hmapclims=(-3, 3), boundsstyle=:dot, :currentscolor=:green)

TopologicalMarkers.plot_auto — Function

plot_auto(<arguments>; <keyword arguments>)

Plots multiple heatmaps, slices or currents simultaneously.

The subplots are automatically arranged into an optimal layout.

Arguments

Each argument can be either a CoordinateRepr object or a chain of pairs.